chore: update with jerels suggestions:
