+++ packages/gatsby-theme-newrelic/gatsby/on-route-update.js
@@ -4,11 +4,10 @@ import getLocale from '../src/utils/getLocale';
 const onRouteUpdate = ({ location, prevLocation }, themeOptions) => {
   trackViaTessen({ location, prevLocation }, themeOptions);
 
-  if (themeOptions.newrelic && window.newrelic) {
-    const { i18n } = themeOptions;
-    const locale = getLocale(location, i18n);
+  if (window.newrelic) {
+    const locale = getLocale(location, themeOptions);
 
-    if (locale !== getLocale(prevLocation, i18n)) {
+    if (locale !== getLocale(prevLocation, themeOptions)) {
       window.newrelic.setCustomAttribute('locale', locale);
     }
   }
