+++ packages/gatsby-theme-newrelic/src/utils/getLocale.js
@@ -1,6 +1,7 @@
-import { defaultLocale } from '../src/utils/defaultOptions';
+import { withDefaults, defaultLocale } from '../src/utils/defaultOptions';
 
-export default getLocale = (i18n, location) => {
+export default getLocale = (themeOptions, location) => {
+  const { i18n } = withDefaults(themeOptions);
   const [, base] = location.pathname.split('/');
   const locale =
     (i18n.additionalLocales || []).find(({ locale }) => locale === base) ||
